Gtonkinson/animations (#225)
* Add signals and scales for opacity animations

These changes add signals and scales for opcity animations for the barSpecBuilder.
As part of these changes, lengendSpecBuilder was also changed to update an existing signal.

* Remove console.warn

* Fix areaUtil tests

* Fix areaSpecBuilder tests

* Update linePointUtils to avoid creating unnecessary keys

* Fix lineMarkUtils tests

* Fix lineSpecBuilder tests

* Fix bar tests

* Fix metricRangeUtils tests

* Add opacity animations to all bar charts

This change allows for opacity animations to all bar charts with a
tooltip or a popover. Animations must be enabled on the chart level
for the animations to appear.

* Refactor code to match style of existing code

* Add opacity animation functionality to line charts.

With this change, all line charts with a tooltip or popover will
animate on hover. Visiually, the animation eases the opacity in and out
on non-hovered lines.

* Add opacity animations to area and trendline.

With these changes, and upon animation enablement, opacity animations
are present with an area, line, and bar chart that has a tooltip, popover,
or trendline.

* Refactor mouseover function to reduce code for opacity animations

* Add opacity animations to legend symbols.

This change includes logic that adds opacity animation to legend
symbols if the current spec has the opacity animation signals for
line, bar, or area charts. This change also has a couple added checks
for interactive children.

* Integrate opacity animations with new changes on main

* Remove unecessary debug settings

* Fix legend hover opacity animations

* Add TODO comments for tracking and testing purposes

* Change opacity animation default

* Change opacity animation gates to expressly check for false

* chore: upgrade storybook to v8

* fix: add testing for animations

* testing for opacity

* fix: add testing for opacity, still need to add opacity story

* fix: add testing to test opacity

* fix opacity stories

* remove readded code

* fix odd indenting

* remove unused def'

---------

Co-authored-by: Adam <adamgrdnr@gmail.com>
Co-authored-by: Jacob <jacobhart1227@gmail.com>
Co-authored-by: Marshall Peterson <marshall.w.peterson@gmail.com>
Co-authored-by: tanugarg04 <61331804+tanugarg04@users.noreply.github.com>