chore: created external and internal BigNumber components and renamed some variables
We were passing rscChartProps to BigNumber so we could render the sparkline using
an rsc chart. There isn't a reason to expose this prop externally though since we'll always
take this prop from the outer Chart component. This creates an external version without the internal
props that returns an empty fragment. We just use this component to grab the props and pass them to
the internal version.